noobschool/requirements.txt
Django==2.1.4
django-filter==2.0.0
django-model-utils==3.1.2
djangorestframework==3.9.0
gunicorn==19.9.0
Markdown==3.0.1
Pillow==5.3.0
psycopg2==2.7.6.1
pytz==2018.7
